requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
twilio==8.10.0
sendgrid==6.11.0
gunicorn==21.2.0
pytz==2024.2

